refactor:  приватные методы Представлений сортировки
