+++ src/controllers/trip.js
@@ -2,8 +2,8 @@ import {sortTypeMap} from '../helpers/const.js';
 import {render, replace, InsertionPosition} from '../helpers/render.js';
 import SortingComponent from '../components/sorting.js';
 import BlankTripComponent from '../components/trip/blank-trip.js';
-import TripWithDaysElement from '../components/trip/trip-with-days.js';
-import TripWithoutDaysElement from '../components/trip/trip-without-days.js';
+import EventsGroupByDaysComponent from '../components/trip/events-group-by-days.js';
+import EventsGroupByTimeOrPriceComponent from '../components/trip/events-group-by-time-or-price.js';
 
 const getTripElement = (sortType, events) => {
   const showingEvents = events.slice();
@@ -13,15 +13,15 @@ const getTripElement = (sortType, events) => {
   switch (sortType) {
     case sortTypeMap.DEFAULT:
       sortedEvents = showingEvents.sort((a, b) => a.date.start - b.date.start);
-      tripElement = new TripWithDaysElement(sortedEvents);
+      tripElement = new EventsGroupByDaysComponent(sortedEvents);
       break;
     case sortTypeMap.TIME:
       sortedEvents = showingEvents.sort((a, b) => (b.date.end - b.date.start) - (a.date.end - a.date.start));
-      tripElement = new TripWithoutDaysElement(sortedEvents);
+      tripElement = new EventsGroupByTimeOrPriceComponent(sortedEvents);
       break;
     case sortTypeMap.PRICE:
       sortedEvents = showingEvents.sort((a, b) => b.price - a.price);
-      tripElement = new TripWithoutDaysElement(sortedEvents);
+      tripElement = new EventsGroupByTimeOrPriceComponent(sortedEvents);
       break;
   }
